+++ tests/milestone0/unit/03_default_config.sh
@@ -8,6 +8,8 @@ OUT=$(feb_stdout -d /tmp/febtest --check-config)
 assert_line "monitor_path=/tmp/febtest"   "$OUT"
 assert_line "socket_path=/tmp/feb.sock"   "$OUT"
 assert_line "recursive=true"              "$OUT"
+assert_line "logical_scope=subtree"       "$OUT"
+assert_contains "FAN_MARK_FILESYSTEM"    "$OUT"
 assert_line "use_io_uring=true"           "$OUT"
 assert_line "log_level=info"              "$OUT"
 assert_line "event_mask=0x8"              "$OUT"